👥 consider all destination aliases of matching image
if I had an emoji 'original' and 'alias_of_original' as an alias in the destination, and tried to transfer 'alias_of_original' again, the transfer_emoji function would only check whether the new name matched 'original', and try to add an alias if not.  Because 'alias_of_original' was taken, it would add 'prefix-alias_of_original'.  This logic checks all aliases of the matching image before deciding to add a new alias.