🏗 separate dictionaries from client
This will let us keep local representations of the destination data to allow us to add multiple emoji at once without making N api calls.

want to use the service as the interface which will call both the dictionary and the client, because otherwise we wouldn't have our existing-destination-emoji references up to date as we add many in a single request.