📂 introduce DirectoryService
this is to allow the transfer CLI to run from a local file of images.

Instead of URLs everywhere, we use relative filepaths.  We introduce
a new method to the image_client to streamline reading local files.

Note that this will not work for the destination yet as much of the
EmojiDict/ImageDict logic for intelligently adding emoji assumes the
emoji paths are URLs.  I don't care much about configuring this
because we have a download script already.